Deleted proof to formalize concepts further, correct errors
There is a mistake in the pdf proving minimum and maximum properties. The concept of a leftmost node is sketchy, the proof sucks to make anything known, there is a terrible assumption that a new leftmost node will heighten the tree when in reality, there can be deeper right nodes (Bad assumption unless only left traversal is considered).